requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.2
aiosignal==1.3.2
alibabacloud-credentials==1.0.2
alibabacloud-credentials-api==1.0.0
alibabacloud-dm20151123==1.2.4
alibabacloud-tea==0.4.3
alibabacloud_endpoint_util==0.0.3
alibabacloud_gateway_spi==0.0.3
alibabacloud_openapi_util==0.2.2
alibabacloud_tea_openapi==0.3.15
alibabacloud_tea_util==0.3.13
alibabacloud_tea_xml==0.0.2
aliyun-python-sdk-core==2.16.0
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
attrs==25.3.0
autopep8==2.3.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cryptography==45.0.3
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.111.0
fastapi-cli==0.0.7
frozenlist==1.6.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.27.0
Hypercorn==0.16.0
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
jmespath==0.10.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.4.4
orjson==3.10.18
outcome==1.3.0.post0
packaging==25.0
pluggy==1.6.0
priority==2.0.0
propcache==0.3.1
pyasn1==0.4.8
pycodestyle==2.13.0
pycparser==2.22
pydantic==2.7.1
pydantic-settings==2.2.1
pydantic_core==2.18.2
Pygments==2.19.1
pyodbc==5.1.0
pytest==8.2.1
pytest-asyncio==0.23.0
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.3
rich==14.0.0
rich-toolkit==0.14.6
rsa==4.9.1
setuptools==78.1.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.37.2
trio==0.24.0
typer==0.15.4
typing_extensions==4.13.2
tzlocal==5.3.1
ujson==5.10.0
urllib3==2.4.0
uvicorn==0.29.0
#uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
wheel==0.45.1
wsproto==1.2.0
yarl==1.20.0
